+++ src/index.tsx
@@ -5,7 +5,14 @@ export {
     fetchConfigFailure,
 }from './js/fetch-config/actions-creators';
 export {fetchConfigReducer} from './js/fetch-config/reducers';
-export {FetchConfigState} from './js/fetch-config/types';
+export {FetchConfigState, FetchConfigAction, FETCH_CONFIG} from './js/fetch-config/types';
+export {
+    userLoggedIn,
+    userLoggedOut
+} from './js/user/actions-creators';
+export {
+    User, USER, UserAction
+} from './js/user/types';
 export {WsBrand} from './js/components/ws-brand/WsBrand';
 export {
     unexpectedErrorOccured,