+++ src/js/fetch-config/reducers.tsx
@@ -1,21 +1,19 @@
 import {FetchConfigAction, FetchConfigState, FETCH_CONFIG} from './types';
 
 const initialState = {
-  isLoading: false,
-  data: null,
+    isLoading: false,
+    data: undefined,
 };
 
 export const fetchConfigReducer = <T extends {}>(
-  state: FetchConfigState<T> = initialState,
-  action: FetchConfigAction<T>): FetchConfigState<T> => {
-  switch (action.type) {
-    case FETCH_CONFIG.REQUEST:
-      return {...state, isLoading: true, data: null};
-    case FETCH_CONFIG.SUCCESS:
-      return {...state, isLoading: false, data: action.data};
-    case FETCH_CONFIG.FAILURE:
-      return {...state, isLoading: false, data: null};
-    default:
-      return state;
-  }
+    state: FetchConfigState<T> = initialState,
+    action: FetchConfigAction<T>): FetchConfigState<T> => {
+    switch (action.type) {
+        case FETCH_CONFIG.REQUEST:
+            return {...state, isLoading: true, data: undefined};
+        case FETCH_CONFIG.FAILURE:
+            return {...state, isLoading: false, data: undefined};
+        default:
+            return state;
+    }
 };