+++ src/js/user/actions-creators.tsx
@@ -1,0 +1,9 @@
+import {User, USER} from "./types";
+
+const userLoggedIn = (user: User) => ({type: USER.LOGGED_IN, data: user});
+const userLoggedOut = () => ({type: USER.LOGGED_OUT});
+
+export {
+    userLoggedIn,
+    userLoggedOut
+}